\name{x00m700p4}
\alias{x00m700p4}
\title{Peak concentration for 100 instantaneous releases}
\description{
  This data set gives the peak concentration for 100 independent
  instantaneous releases of neutral-buoyancy gas in a windtunnel
}
\usage{data(x00m700p4)}
\format{A vector containing 100 observations}
\references{
  D. J. Hall and others 1991.  \emph{Repeat variability in
    instantaneously released heavy gas clouds---some wind tunnel model
    experiments}.  Technical Report LR 804 (PA), Warren Spring
    Laboratory, Gunnels Wood Road, Stevenage, Hertfordshire SG1 2BX.
}
\examples{
data(x00m700p4)
fit.davies.q(x00m700p4)
}

\keyword{datasets}
